refactor(chat): 提取展示叠加层
